Conflicts:
	scripts/generate_otu_signifigance_tables_AGP.py

+++ americangut/generate_otu_signifigance_tables.py
@@ -3,7 +3,6 @@
 from numpy import loadtxt, delete, mean, shape, argsort, sort
 from scipy.stats import ttest_1samp
 
-
 __author__ = "Justine Debelius"
 __copyright__ = "Copyright 2013, The American Gut Project"
 __credits__ = ["Justine Debelius"]
@@ -107,9 +106,11 @@ def calculate_tax_rank_1(sample, population, taxa, rare_threshold=0.1, \
     # Identifies the taxonomy up to the abundance threshold    
     abundance_watch = 0
     abundant = []
+
     for idx, frequency in enumerate(reversed(abundance_data)):
+        tax_idx = len(abundance_data) - (idx + 1)
         abundance_watch = abundance_watch + frequency
-        abundant.append([abundance_taxa[idx], frequency])
+        abundant.append([abundance_taxa[tax_idx], frequency])
         if abundance_watch > abundance_threshold:
             break
 
@@ -126,6 +127,11 @@ def calculate_tax_rank_1(sample, population, taxa, rare_threshold=0.1, \
 
         elif sample_bin[idx] == 1 and \
              population_count[idx] < num_samples*rare_threshold:
+
+            # ignore contested groupings
+            if '[' in taxon:
+                continue
+
             rare.append(taxon)
             remove_index.append(idx)
 
@@ -606,4 +612,3 @@ def convert_taxa_to_table(corr_taxa, header, render_mode = "RAW", \
     table = anterow.join(table_code)
 
     return table
-
